Fix: move packet ports from domain/packet/core to their consumers
The core package held adapter ports, not domain concepts, and its name said
nothing. Go's implicit interfaces let each port live with the code that calls
it, so domain/ is left holding only types and rules.

- IpFilter    -> application/lists
- Reader      -> application/packetstream
- EbpfLoader  -> application/packet
- domain/packet/core removed
