+++ source/application/lists/service.go
@@ -1,15 +1,14 @@
 package lists
 
 import (
-	coreDomain "ntc/source/domain/packet/core"
 	coreNetwork "ntc/source/domain/network/core"
 )
 
 type ListService struct {
-	filter coreDomain.IpFilter
+	filter     IpFilter
 	cidrFilter coreNetwork.CIDRFilter
 }
 
-func NewListService(f coreDomain.IpFilter, cidrFilter coreNetwork.CIDRFilter) *ListService {
+func NewListService(f IpFilter, cidrFilter coreNetwork.CIDRFilter) *ListService {
 	return &ListService{filter: f, cidrFilter: cidrFilter}
 }
