+++ source/application/packet/loader.go
@@ -1,0 +1,19 @@
+package packet
+
+import (
+	"context"
+	"ntc/source/application/lists"
+	"ntc/source/application/packetstream"
+	"ntc/source/domain/network/core"
+)
+
+type EbpfLoader interface {
+	NewIpFilter() lists.IpFilter
+	NewCIDRFilter() core.CIDRFilter
+
+	LoadCollection(objPath string) error
+	ListenInterface(interfaceName string) error
+	AttachProgram() error
+	NewReader(ctx context.Context) (packetstream.Reader, error)
+	Close() error
+}
